Document modular design preference
